refactor util tests
